[#3] Refactoring and renaming methods in common.
