+++ morphman/manipulate_bend.py
@@ -101,7 +101,7 @@ def manipulate_bend(input_filepath, output_filepath, smooth, smooth_factor, regi
 
     # Clip centerline
     print("-- Clipping centerlines.")
-    locator = get_locator(extract_single_line(centerlines, 0))
+    locator = get_vtk_point_locator(extract_single_line(centerlines, 0))
     id1 = locator.FindClosestPoint(region_points[0])
     id2 = locator.FindClosestPoint(region_points[1])
     p1 = centerlines.GetPoint(id1)
@@ -112,7 +112,7 @@ def manipulate_bend(input_filepath, output_filepath, smooth, smooth_factor, regi
 
     if diverging_centerline_ispresent:
         diverging_centerline_end = extract_single_line(patch_diverging_line, 1)
-        centerline_bend = merge_data([centerline_bend, diverging_centerline_end])
+        centerline_bend = vtk_append_polydata([centerline_bend, diverging_centerline_end])
 
     write_polydata(centerline_remaining, centerline_clipped_path)
     write_polydata(centerline_bend, centerline_clipped_part_path)
@@ -151,7 +151,7 @@ def manipulate_bend(input_filepath, output_filepath, smooth, smooth_factor, regi
                                                  diverging_centerline_ispresent=diverging_centerline_ispresent)
     else:
         if diverging_centerline_ispresent:
-            new_centerlines = merge_data([centerlines, extract_single_line(diverging_centerlines, 0)])
+            new_centerlines = vtk_append_polydata([centerlines, extract_single_line(diverging_centerlines, 0)])
         else:
             new_centerlines = centerlines
 
@@ -160,12 +160,12 @@ def manipulate_bend(input_filepath, output_filepath, smooth, smooth_factor, regi
 
     if alpha == 0.0 and beta != 0.0:
         print("-- Creating new surface.")
-        new_voronoi = merge_data([voronoi_remaining, voronoi_bend])
+        new_voronoi = vtk_append_polydata([voronoi_remaining, voronoi_bend])
         new_surface = create_new_surface(new_voronoi, poly_ball_size=poly_ball_size)
 
     elif alpha != 0.0:
         # Update the region points
-        locator = get_locator(new_centerlines)
+        locator = get_vtk_point_locator(new_centerlines)
         region_points[0] = new_centerlines.GetPoint(locator.FindClosestPoint(region_points[0]))
         region_points[1] = new_centerlines.GetPoint(locator.FindClosestPoint(region_points[1]))
 
@@ -179,8 +179,8 @@ def manipulate_bend(input_filepath, output_filepath, smooth, smooth_factor, regi
     new_surface = prepare_surface_output(new_surface, surface,
                                          new_centerlines, output_filepath,
                                          test_merge=True, changed=True,
-                                         old_centerline=merge_data([centerlines,
-                                                                    diverging_centerlines]))
+                                         old_centerline=vtk_append_polydata([centerlines,
+                                                                             diverging_centerlines]))
     write_polydata(new_centerlines, new_centerlines_path)
     write_polydata(new_surface, output_filepath)
 
@@ -218,7 +218,7 @@ def manipulate_bend_vertically(alpha, voronoi_remaining, voronoi_bend, centerlin
 
     # Get clipped curve
     print("-- Clipping centerlines.")
-    locator = get_locator(extract_single_line(centerlines, 0))
+    locator = get_vtk_point_locator(extract_single_line(centerlines, 0))
     id1 = locator.FindClosestPoint(region_points[0])
     id2 = locator.FindClosestPoint(region_points[1])
     p1 = centerlines.GetPoint(id1)
@@ -227,7 +227,7 @@ def manipulate_bend_vertically(alpha, voronoi_remaining, voronoi_bend, centerlin
 
     if diverging_centerline_ispresent:
         diverging_centerline_end = extract_single_line(patch_diverging_line, 1)
-        centerline_bend = merge_data([centerline_bend, diverging_centerline_end])
+        centerline_bend = vtk_append_polydata([centerline_bend, diverging_centerline_end])
 
     # Find ID of middle point:
     print("-- Finding points to spline through.")
@@ -239,7 +239,7 @@ def manipulate_bend_vertically(alpha, voronoi_remaining, voronoi_bend, centerlin
     print("-- Adjust Voronoi diagram")
     voronoi_bend = move_voronoi_vertically(voronoi_bend, centerline_bend, id1,
                                            diverging_id, dx, diverging_centerline_ispresent)
-    new_voronoi = merge_data([voronoi_remaining, voronoi_bend])
+    new_voronoi = vtk_append_polydata([voronoi_remaining, voronoi_bend])
 
     # Move centerline manually for postprocessing
     new_centerlines = move_centerlines(centerlines, dx, p1, p2, diverging_id, diverging_centerlines, direction)
@@ -272,7 +272,7 @@ def move_voronoi_horizontally(dx_p1, voronoi_clipped, centerline_clipped, id1, i
         new_dataset (vtkPolyData): Manipulated Voronoi diagram.
     """
 
-    centerline_loc = get_locator(centerline_clipped)
+    centerline_loc = get_vtk_point_locator(centerline_clipped)
     new_dataset = vtk.vtkPolyData()
     points = vtk.vtkPoints()
     verts = vtk.vtkCellArray()
@@ -369,7 +369,7 @@ def move_voronoi_vertically(voronoi_clipped, centerline_clipped, id1_0, clip_id,
         new_dataset (vtkPolyData): Manipulated Voronoi diagram.
     """
 
-    centerline_loc = get_locator(centerline_clipped)
+    centerline_loc = get_vtk_point_locator(centerline_clipped)
     new_dataset = vtk.vtkPolyData()
     points = vtk.vtkPoints()
     verts = vtk.vtkCellArray()
