+++ morphman/misc/automated_landmarking.py
@@ -56,7 +56,7 @@ def map_landmarks(landmarks, centerline):
     Returns:
         landmarks (dict): Contains landmarks mapped to centerline.
     """
-    locator = get_locator(centerline)
+    locator = get_vtk_point_locator(centerline)
     for key in landmarks:
         landmark = landmarks[key]
         landmark_id = locator.FindClosestPoint(landmark)
@@ -96,7 +96,7 @@ def landmarking_bogunovic(centerline, base_path, curv_method, algorithm,
     if curv_method == "vmtk":
         line = vmtk_centerline_attributes(centerline)
         line = vmtk_centerline_geometry(line, smooth_line, factor=smoothing_factor, iterations=iterations)
-        curvature_ = get_array("Curvature", line)
+        curvature_ = get_point_data_array("Curvature", line)
         curvature__ = gaussian_filter(curvature_, 5)
         curvature = []
         for c in curvature__:
@@ -115,7 +115,7 @@ def landmarking_bogunovic(centerline, base_path, curv_method, algorithm,
 
     elif curv_method == "spline":
         line, max_point_ids, min_point_ids = spline_and_geometry(centerline, smooth_line, nknots)
-        curvature = get_array("Curvature", line)
+        curvature = get_point_data_array("Curvature", line)
 
     if curv_method != "spline":
         max_point_ids = list(argrelextrema(curvature, np.greater)[0])
@@ -123,8 +123,8 @@ def landmarking_bogunovic(centerline, base_path, curv_method, algorithm,
         get_k1k2_basis(curvature, line)
 
     length = get_curvilinear_coordinate(line)
-    k1 = get_array("k1", line)
-    k2 = get_array("k2", line)
+    k1 = get_point_data_array("k1", line)
+    k2 = get_point_data_array("k2", line)
 
     # Remove a min / max point that is in reality a saddle point
     for i in min_point_ids:
@@ -273,8 +273,8 @@ def landmarking_piccinelli(centerline, base_path, curv_method, algorithm, resamp
         line, max_point_ids, min_point_ids = spline_and_geometry(centerline, smooth_line, nknots)
 
         # Get curvature and torsion, find peaks
-        curvature = get_array("Curvature", line)
-        torsion = get_array("Torsion", line)
+        curvature = get_point_data_array("Curvature", line)
+        torsion = get_point_data_array("Torsion", line)
         torsion_smooth = gaussian_filter(torsion, 10)
         max_point_tor_ids = list(argrelextrema(abs(torsion_smooth), np.greater)[0])
 
@@ -284,8 +284,8 @@ def landmarking_piccinelli(centerline, base_path, curv_method, algorithm, resamp
         line_tor = vmtk_centerline_geometry(centerline, True, outputsmoothed=False,
                                             factor=smoothing_factor_torsion, iterations=iterations)
         # Get curvature and torsion, find peaks
-        curvature = get_array("Curvature", line)
-        torsion = get_array("Torsion", line_tor)
+        curvature = get_point_data_array("Curvature", line)
+        torsion = get_point_data_array("Torsion", line_tor)
         torsion_smooth = gaussian_filter(torsion, 10)
         curvature_smooth = gaussian_filter(curvature, 10)
         max_point_ids = list(argrelextrema(curvature_smooth, np.greater)[0])
@@ -410,7 +410,7 @@ def spline_and_geometry(line, smooth, nknots):
     data[:, 2] = fz_
 
     header = ["X", "Y", "Z"]
-    line = data_to_vtkPolyData(data, header)
+    line = data_to_vtk_polydata(data, header)
 
     # Let vmtk compute curve attributes
     line = vmtk_centerline_attributes(line)
@@ -436,7 +436,7 @@ def spline_and_geometry(line, smooth, nknots):
     max_point_ids = list(argrelextrema(curvature_, np.greater)[0])
     min_point_ids = list(argrelextrema(curvature_, np.less)[0])
 
-    locator = get_locator(line)
+    locator = get_vtk_point_locator(line)
 
     min_points = [[fx_[i], fy_[i], fz_[i]] for i in min_point_ids]
     max_points = [[fx_[i], fy_[i], fz_[i]] for i in max_point_ids]
@@ -447,7 +447,7 @@ def spline_and_geometry(line, smooth, nknots):
         min_point_ids.append(locator.FindClosestPoint(point_min))
         max_point_ids.append(locator.FindClosestPoint(point_max))
 
-    curvature = get_array("Curvature", line)
+    curvature = get_point_data_array("Curvature", line)
     line = get_k1k2_basis(curvature, line)
 
     length = get_curvilinear_coordinate(line)
