+++ test/test_manipulate_bifurcation.py
@@ -9,8 +9,8 @@ import pytest
 import numpy as np
 
 from .fixtures import common_input
-from morphman.common import get_path_names, read_polydata, get_locator, get_tolerance, \
-                            extract_single_line, distance
+from morphman.common import get_path_names, read_polydata, get_vtk_point_locator, get_tolerance, \
+                            extract_single_line, get_distance
 from morphman import manipulate_bifurcation
 
 
@@ -48,8 +48,8 @@ def test_bifurcation_angle(common_input, angle):
         line_old = extract_single_line(old_centerlines, i)
         line_new = extract_single_line(new_centerlines, i)
 
-        loc_old = get_locator(line_old)
-        loc_new = get_locator(line_new)
+        loc_old = get_vtk_point_locator(line_old)
+        loc_new = get_vtk_point_locator(line_new)
 
         id1_old = loc_old.FindClosestPoint(start_point1)
         id2_old = loc_old.FindClosestPoint(start_point2)
@@ -57,16 +57,16 @@ def test_bifurcation_angle(common_input, angle):
         id1_new = loc_new.FindClosestPoint(start_point1)
         id2_new = loc_new.FindClosestPoint(start_point2)
 
-        if distance(start_point1, line_old.GetPoint(id1_old)) < tol:
+        if get_distance(start_point1, line_old.GetPoint(id1_old)) < tol:
             cl_old_1 = i
             cl_old_id1 = id1_old
-        if distance(start_point2, line_old.GetPoint(id2_old)) < tol:
+        if get_distance(start_point2, line_old.GetPoint(id2_old)) < tol:
             cl_old_2 = i
             cl_old_id2 = id2_old
-        if distance(start_point1, line_new.GetPoint(id1_new)) < tol:
+        if get_distance(start_point1, line_new.GetPoint(id1_new)) < tol:
             cl_new_1 = i
             cl_new_id1 = id1_new
-        if distance(start_point2, line_new.GetPoint(id2_new)) < tol:
+        if get_distance(start_point2, line_new.GetPoint(id2_new)) < tol:
             cl_new_2 = i
             cl_new_id2 = id2_new
